% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{aggregate_to_clusters}
\alias{aggregate_to_clusters}
\title{Aggregate AOI-level data to cluster-level for Stage B}
\usage{
aggregate_to_clusters(
  prep,
  clusters,
  tech_spatial,
  group_col = "sim_group",
  aggregate_tech = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{prep}{Output of prep_inla_long_df()}

\item{clusters}{Output of make_aoi_clusters()}

\item{tech_spatial}{Numeric vector of length n_AOI (Stage A spatial effect)}

\item{verbose}{Logical}

\item{diff_mode}{How to aggregate diff covariate ("mean" or "majority")}
}
\value{
list(df = aggregated_df, aoi = aggregated_aoi)
}
\description{
Aggregate AOI-level data to cluster-level for Stage B
}
