% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_inla_baseline_model}
\alias{fit_inla_baseline_model}
\title{Baseline INLA normalization model: gene baseline + spatial AOI}
\usage{
fit_inla_baseline_model(
  prep,
  k = 4,
  symmetric = TRUE,
  family = "nbinomial",
  verbose = FALSE
)
}
\description{
Baseline INLA normalization model: gene baseline + spatial AOI
}
