% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_inla_mix}
\alias{fit_inla_mix}
\title{Baseline INLA model with two-component DE mixture prior}
\usage{
fit_inla_mix(
  prep,
  k = 4,
  symmetric = TRUE,
  family = "nbinomial",
  verbose = FALSE,
  small_U = 0.1,
  small_alpha = 0.5,
  large_U = 0.75,
  large_alpha = 0.01
)
}
\description{
Heavy-tailed approximation:
beta_de = beta_small + beta_large
}
