% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{normalize_spe_inla_pac}
\alias{normalize_spe_inla_pac}
\title{Normalize a SpatialExperiment using an INLA + PAC model}
\usage{
normalize_spe_inla_pac(
  spe,
  model_fun,
  group_col = "sim_group",
  counts_assay = "counts",
  tech_covariates = NULL,
  center_scale_covariates = FALSE,
  subsample_frac = NULL,
  model_args = list(),
  cluster_aoi = FALSE,
  cluster_args = list()
)
}
\value{
List with SpatialExperiment (\code{spe}) and fitted model (\code{fit})
}
\description{
Normalize a SpatialExperiment using an INLA + PAC model
}
