% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{spe_to_geomx_long}
\alias{spe_to_geomx_long}
\title{Convert SpatialExperiment to GeoMx long-form tibble}
\usage{
spe_to_geomx_long(spe, assay = "counts", group_col = "Mix")
}
\arguments{
\item{spe}{SpatialExperiment.}

\item{assay}{Name of assay to extract (default "counts").}

\item{group_col}{Name of group variable in colData to include (default "Mix").}
}
\value{
Long-form tibble with AOI_ID, Gene_ID, Count_int and metadata.
}
\description{
Convert SpatialExperiment to GeoMx long-form tibble
}
