% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_spe_by_slide}
\alias{split_spe_by_slide}
\title{Split a simulated SpatialExperiment into a list of per-slide objects}
\usage{
split_spe_by_slide(spe, slide_col = "slide_id")
}
\arguments{
\item{spe}{A SpatialExperiment produced by simulate_spe_benchmark().}

\item{slide_col}{Character. Column in colData(spe) holding slide IDs.
Default "slide_id".}
}
\value{
A named list of SpatialExperiment objects, one per slide.
}
\description{
Split a simulated SpatialExperiment into a list of per-slide objects
}
