% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{validate_geomx_long}
\alias{validate_geomx_long}
\title{Validate required columns in a GeoMx long-form tibble}
\usage{
validate_geomx_long(data, required)
}
\arguments{
\item{data}{A data.frame/tibble in long format.}

\item{required}{Character vector of required column names.}
}
\value{
Invisibly returns TRUE on success; errors otherwise.
}
\description{
Validate required columns in a GeoMx long-form tibble
}
\keyword{internal}
